projectiles will now become a static object at the conclusion of projectile motion. heading is nolonger an sbyte.

+++ src/conversation/conversation_functions/x_obj_stuff.cs
@@ -68,7 +68,7 @@ namespace Underworld
                     {
                         if (look.CanBeIdentified(obj))
                         {
-                            obj.heading = (ushort)identifyObject;
+                            obj.heading = identifyObject;
                         }
                     }
                     if (owner!=-1)
